+++ app/src/main/java/com/nike/snkrs/sunderland/ui/view/FragmentSplashScreen.kt
@@ -14,12 +14,14 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.annotation.VisibleForTesting
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import com.nike.snkrs.sunderland.R
 import com.nike.snkrs.sunderland.util.tryCatchWithLogging
 import kotlinx.coroutines.*
-import java.util.concurrent.TimeUnit
+import kotlin.time.Duration
+import kotlin.time.ExperimentalTime
 
 //endregion import directives
 
@@ -28,6 +30,7 @@ import java.util.concurrent.TimeUnit
  * Used to display a "splash" screen prior to showing the main content
  * @author Thomas Sunderland. 2021 MAY 09
  */
+@ExperimentalTime
 class FragmentSplashScreen : Fragment(), CoroutineScope {
 
     //region companion object
@@ -39,7 +42,7 @@ class FragmentSplashScreen : Fragment(), CoroutineScope {
         /**
          * Number of seconds to display the splash screen before proceeding to the next screen
          */
-        private const val SPLASH_SCREEN_DURATION_SECONDS = 3L
+        private val SPLASH_SCREEN_DURATION_SECONDS = Duration.seconds(3)
         //endregion constants
     }
     //endregion companion object
@@ -82,30 +85,6 @@ class FragmentSplashScreen : Fragment(), CoroutineScope {
     //@formatter:on
 
     /**
-     * Called immediately after {@link #onCreateView(LayoutInflater, ViewGroup, Bundle)} has returned,
-     * but before any saved state has been restored in to the view. This gives subclasses a chance to
-     * initialize themselves once they know their view hierarchy has been completely created. The fragment's
-     * view hierarchy is not however attached to its parent at this point.
-     *
-     * @param view The View returned by {@link #onCreateView(LayoutInflater, ViewGroup, Bundle)}.
-     * @param savedInstanceState If non-null, this fragment is being re-constructed from a previous saved state as given here.
-     */
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        tryCatchWithLogging({
-            // call into base class implementation
-            super.onViewCreated(view, savedInstanceState)
-
-            // wait a bit and then navigate to the sneakers fragment
-            navigateAfterDelayRoutine = launch {
-                delay(TimeUnit.SECONDS.toMillis(SPLASH_SCREEN_DURATION_SECONDS))
-                withContext(Dispatchers.Main) {
-                    findNavController().navigate(FragmentSplashScreenDirections.actionFragmentSplashScreenToFragmentSneakers())
-                }
-            }
-        })
-    }
-
-    /**
      * Called when the Fragment is visible to the user.
      */
     override fun onStart() {
@@ -118,6 +97,11 @@ class FragmentSplashScreen : Fragment(), CoroutineScope {
                 supportActionBar?.hide()
                 viewBinding.bottomNavigationView.visibility = View.GONE
             }
+
+            // wait a bit and then navigate to the sneakers fragment
+            navigateAfterDelayRoutine = launch {
+                screenTransitionWithDelay(SPLASH_SCREEN_DURATION_SECONDS)
+            }
         })
     }
 
@@ -129,6 +113,10 @@ class FragmentSplashScreen : Fragment(), CoroutineScope {
             // call into base class implementation
             super.onStop()
 
+            // cancel the navigation routine
+            // note: this effectively handles device rotation by preventing navigation from a destroyed context
+            if (::navigateAfterDelayRoutine.isInitialized) navigateAfterDelayRoutine.cancel()
+
             // show app bar, bottom navigation on stop
             with(activity as ActivityMain) {
                 supportActionBar?.show()
@@ -136,20 +124,22 @@ class FragmentSplashScreen : Fragment(), CoroutineScope {
             }
         })
     }
+    //endregion lifecycle overrides
+
+
+    //region functions
 
     /**
-     * Called when the fragment is no longer in use. This is called
-     * after {@link #onStop()} and before {@link #onDetach()}.
+     * Delays for the specified duration and then auto-navigates to the sneakers fragment
      */
-    override fun onDestroy() {
+    @VisibleForTesting
+    suspend fun screenTransitionWithDelay(delay: Duration) {
         tryCatchWithLogging({
-            // call into base class implementation
-            super.onDestroy()
-
-            // cancel the navigation routine
-            // note: this effectively handles device rotation by preventing navigation from a destroyed context
-            if (::navigateAfterDelayRoutine.isInitialized) navigateAfterDelayRoutine.cancel()
-        })
+            delay(delay.inWholeMilliseconds)
+            withContext(Dispatchers.Main) {
+                findNavController().navigate(FragmentSplashScreenDirections.actionFragmentSplashScreenToFragmentSneakers())
+            }
+        }, listOf(delay))
     }
-    //endregion lifecycle overrides
+    //endregion functions
 }