+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 #CHANGELOG- Sunderland Coding Assessment
 
-# Version 1.0.0 Release Notes
+# Version 1.21.6.0 Release Notes
+  - Improvement - Updated the splash screen transition delay to use the type-safe Duration class as opposed to Long delay, broke delay logic out into a more easily testable function and added test cases
+  - Bug Fix - Splash screen gets stuck if the screen is turned off while the coroutine is running
+  - General - Updated dependencies
+
+
+# Version 1.21.5.0 Release Notes
   - Baseline release